+++ frontend/src/components/Header/test.js
@@ -1,31 +1,12 @@
 import React from 'react';
-import { RecoilRoot } from 'recoil';
-import { ThemeProvider } from 'styled-components';
-import { render, screen, fireEvent } from '@testing-library/react';
+import { render, screen, fireEvent } from '../../test-utils';
 import '@testing-library/jest-dom';
 
-import theme from '../../styles/theme';
 import Header from '.';
-
-import { BrowserRouter } from 'react-router-dom';
-
-const AllTheProviders = ({ children }) => {
-    return (
-        <RecoilRoot>
-            <ThemeProvider theme={theme}>
-                <BrowserRouter>{children}</BrowserRouter>
-            </ThemeProvider>
-        </RecoilRoot>
-    );
-};
-
-const customRender = (ui, options) => {
-    return render(ui, { wrapper: AllTheProviders, ...options });
-};
 
 describe('공통 헤더 컴포넌트 테스트', () => {
     it('이전 달, 다음 달로 이동할 수 있는 Arrow 버튼이 2개 표시된다.', () => {
-        customRender(<Header current="main" />);
+        render(<Header current="main" />);
 
         const arrowButtons = screen.getAllByRole('button', { name: /month/i });
 
@@ -33,7 +14,7 @@ describe('공통 헤더 컴포넌트 테스트', () => {
     });
 
     it('초기 렌더링 시 날짜는 현재 날짜의 년, 월이 표시된다.', () => {
-        customRender(<Header current="main" />);
+        render(<Header current="main" />);
 
         const year = new Date().getFullYear();
         const month = new Date().getMonth() + 1;
@@ -43,7 +24,7 @@ describe('공통 헤더 컴포넌트 테스트', () => {
     });
 
     it('내역, 달력, 통계로 이동할 수 있는 Link 요소가 3개 표시된다.', () => {
-        customRender(<Header current="main" />);
+        render(<Header current="main" />);
 
         const arrowButtons = screen.getAllByRole('link');
 
@@ -51,7 +32,7 @@ describe('공통 헤더 컴포넌트 테스트', () => {
     });
 
     it('이전 달로 이동하는 Arrow 버튼 클릭 시 날짜는 이전 달로 표시된다.', () => {
-        customRender(<Header current="main" />);
+        render(<Header current="main" />);
 
         const dateStrong = screen.getByTestId('date');
         const [year, month] = dateStrong.textContent.split(/[^0-9]+/g);
@@ -65,7 +46,7 @@ describe('공통 헤더 컴포넌트 테스트', () => {
     });
 
     it('다음 달로 이동하는 Arrow 버튼 클릭 시 날짜는 다음 달로 표시된다.', () => {
-        customRender(<Header current="main" />);
+        render(<Header current="main" />);
 
         const dateStrong = screen.getByTestId('date');
         const [year, month] = dateStrong.textContent.split(/[^0-9]+/g);
