+++ frontend/src/components/Summary/test.js
@@ -1,29 +1,12 @@
-import React, { Suspense } from 'react';
-import { RecoilRoot } from 'recoil';
-import { ThemeProvider } from 'styled-components';
-import { render, screen, fireEvent } from '@testing-library/react';
+import React from 'react';
+import { render, screen, fireEvent } from '../../test-utils';
 import '@testing-library/jest-dom';
 
-import theme from '../../styles/theme';
 import Summary from '.';
-
-const AllTheProviders = ({ children }) => {
-    return (
-        <RecoilRoot>
-            <ThemeProvider theme={theme}>
-                <Suspense fallback={<div>Loading...</div>}>{children}</Suspense>
-            </ThemeProvider>
-        </RecoilRoot>
-    );
-};
-
-const customRender = (ui, options) => {
-    return render(ui, { wrapper: AllTheProviders, ...options });
-};
 
 describe('메인 화면(지출입 내역) Summary 컴포넌트 테스트', () => {
     it('총 이체, 수입, 지출까지 3개의 체크박스가 표시된다.', async () => {
-        customRender(<Summary />);
+        render(<Summary />);
 
         const checkBoxes = await screen.findAllByRole('checkbox');
 
@@ -31,7 +14,7 @@ describe('메인 화면(지출입 내역) Summary 컴포넌트 테스트', () =>
     });
 
     it('현재 날짜에 해당하는 총 이체 건수가 정상적으로 표시된다.', async () => {
-        customRender(<Summary />);
+        render(<Summary />);
 
         const totalCountSpan = await screen.findByTestId('totalCount');
 
@@ -39,7 +22,7 @@ describe('메인 화면(지출입 내역) Summary 컴포넌트 테스트', () =>
     });
 
     it('현재 날짜에 해당하는 수입이 정상적으로 표시된다.', async () => {
-        customRender(<Summary />);
+        render(<Summary />);
 
         const incomeSpan = await screen.findByTestId('income');
 
@@ -47,7 +30,7 @@ describe('메인 화면(지출입 내역) Summary 컴포넌트 테스트', () =>
     });
 
     it('현재 날짜에 해당하는 지출이 정상적으로 표시된다.', async () => {
-        customRender(<Summary />);
+        render(<Summary />);
 
         const expenditureSpan = await screen.findByTestId('expenditure');
